error popups

+++ src/gui/controller/MainViewController.java
@@ -23,7 +23,9 @@ import java.io.File;
 import java.io.IOException;
 import java.net.URL;
 import java.sql.SQLException;
+import java.util.ArrayList;
 import java.util.Objects;
+import java.util.Optional;
 import java.util.ResourceBundle;
 
 public class MainViewController implements Initializable {
@@ -99,7 +101,6 @@ public class MainViewController implements Initializable {
         moviesLastViewCol.setCellValueFactory(new PropertyValueFactory<Movie, String>("LastView"));
 
 
-
         try {
             allMoviesTable.setItems(movieModel.getObservableMovies());
             categoryListView.setItems(categoryModel.getObservableCategories());
@@ -111,7 +112,7 @@ public class MainViewController implements Initializable {
         filterType = "movieFilter";
         filterInput.textProperty().addListener((observableValue, oldValue, newValue) -> {
             try {
-                movieModel.searchMovie(newValue,filterType);
+                movieModel.searchMovie(newValue, filterType);
             } catch (Exception e) {
                 e.printStackTrace();
             }
@@ -125,10 +126,10 @@ public class MainViewController implements Initializable {
 
     /**
      * Metoden handleCreateCategory styrer knappen "Create" under Category, som åbner createCategory.fxml
+     *
      * @param actionEvent javaFX event klasse
      * @throws IOException
      */
-
     public void handleCreateCategory(ActionEvent actionEvent) throws IOException {
         Parent parent = FXMLLoader.load(Objects.requireNonNull(getClass().getResource("/gui/view/CreateCategory.fxml"))); // The FXML path
         Scene scene = new Scene(parent); // Scenen som skal vises
@@ -142,21 +143,16 @@ public class MainViewController implements Initializable {
 
     /**
      * Metoden handleEditCategory styrer knappen "Edit" under Category, som åbner EditCategory.fxml
+     *
      * @param actionEvent javaFX event klasse
      * @throws IOException
      */
-
     public void handleEditCategory(ActionEvent actionEvent) throws IOException {
         Category selectedCategory = categoryListView.getSelectionModel().getSelectedItem();
         if (selectedCategory != null) {
             FXMLLoader root = new FXMLLoader(getClass().getResource("/gui/view/EditCategory.fxml"));
-            Scene mainWindowScene = null;
+            Scene mainWindowScene = new Scene(root.load());
 
-            try {
-                mainWindowScene = new Scene(root.load());
-            } catch (IOException ioException) {
-                ioException.printStackTrace();
-            }
             Stage editCategoryStage = new Stage();
             editCategoryStage.setScene(mainWindowScene);
             editCategoryStage.setResizable(false);
@@ -166,29 +162,37 @@ public class MainViewController implements Initializable {
             categoryListView.getItems().clear();
             categoryListView.setItems(categoryModel.getObservableCategories());
         }
+        else{
+            error("Select a category and try again");
+        }
     }
 
     /**
      * handleDeleteCategory styrer knappen "delete" under Category listView. Metoden sletter et specifikt valgt category objekt
+     *
      * @param actionEvent javaFX event klasse
      */
-
     public void handleDeleteCategory(ActionEvent actionEvent) {
         if (categoryListView.getSelectionModel().getSelectedItem() == null) {
             error("Please choose a category to delete");
         } else {
-            categoryModel.deleteCategory(categoryListView.getSelectionModel().getSelectedItem());
-            categoryListView.getItems().remove(categoryListView.getSelectionModel().getSelectedItem());
+            Category category = categoryListView.getSelectionModel().getSelectedItem();
+            Alert alert = new Alert(Alert.AlertType.CONFIRMATION, "Are you sure you want to delete "+category.getName(), ButtonType.YES, ButtonType.NO);
+            Optional<ButtonType> result = alert.showAndWait();
+            if (result.isPresent() && result.get() == ButtonType.YES) {
+                categoryModel.deleteCategory(categoryListView.getSelectionModel().getSelectedItem());
+                categoryListView.getItems().remove(categoryListView.getSelectionModel().getSelectedItem());
+                moviesTable.getItems().clear();
+            }
         }
     }
 
-
     /**
      * handleCreateMovieButton styrer knappen "Create" under allMovies tableview, som åbner "gui/view/CreateMovie.fxml"
+     *
      * @param actionEvent javaFX event klasse
      * @throws IOException
      */
-
     public void handleCreateMovieButton(ActionEvent actionEvent) throws IOException {
         Parent root = FXMLLoader.load(getClass().getResource("/gui/view/CreateMovie.fxml"));
         Stage stage = new Stage();
@@ -202,10 +206,10 @@ public class MainViewController implements Initializable {
 
     /**
      * handleEditMovieButton styrer knappen "Edit" under allMovies tableview, som åbner "gui/view/EditMovie.fxml"
+     *
      * @param actionEvent javaFX event klasse
      * @throws IOException
      */
-
     public void handleEditMovieButton(ActionEvent actionEvent) throws IOException {
         Movie selectedMovie = allMoviesTable.getSelectionModel().getSelectedItem();
         if (selectedMovie != null) {
@@ -220,28 +224,35 @@ public class MainViewController implements Initializable {
             editMovieStage.showAndWait();
             allMoviesTable.getItems().clear();
             allMoviesTable.setItems(movieModel.getObservableMovies());
+        }else {
+            error("Select a movie and try again");
         }
     }
 
     /**
-     * handleDeleteMovieButton styrer knappen "delete" under Movie TableView. Metoden sletter et specifikt valgt Movie objekt
+     * handleDeleteMovieButton styrer knappen "delete" under Movie TableView. Metoden sletter et specifikt valgt Movie objekt hvis der vælges yes i alertboxen
+     *
      * @param actionEvent javaFX event klasse
      */
-
     public void handleDeleteMovieButton(ActionEvent actionEvent) {
         if (allMoviesTable.getSelectionModel().getSelectedItem() == null) {
             error("Please choose a movie to delete");
         } else {
-            movieModel.deleteMovie(allMoviesTable.getSelectionModel().getSelectedItem());
-            allMoviesTable.getItems().remove(allMoviesTable.getSelectionModel().getSelectedItem());
+            Alert alert = new Alert(Alert.AlertType.CONFIRMATION, "Are you sure you want to delete?", ButtonType.YES, ButtonType.NO);
+            Optional<ButtonType> result = alert.showAndWait();
+            if (result.isPresent() && result.get() == ButtonType.YES) {
+                Movie movie = allMoviesTable.getSelectionModel().getSelectedItem();
+                movieModel.deleteMovie(movie);
+                allMoviesTable.getItems().remove(movie);
+            }
         }
     }
 
     /**
      * Error message
+     *
      * @param text
      */
-
     private void error(String text) {
         Alert alert = new Alert(Alert.AlertType.ERROR, text, ButtonType.OK);
         alert.showAndWait();
@@ -249,36 +260,56 @@ public class MainViewController implements Initializable {
 
     /**
      * addMovieToCategoryBtn Styrer knappen "Add", som tilføjer et Movie objekt til en Category objekt
+     *
      * @param actionEvent javaFX event klasse
      */
     public void addMovieToCategoryBtn(ActionEvent actionEvent) {
+        boolean alReadyInCategory = false;
         Category selectedCategory = categoryListView.getSelectionModel().getSelectedItem();
         Movie selectedMovie = allMoviesTable.getSelectionModel().getSelectedItem();
-        categoryModel.addMovieToCategory(selectedCategory, selectedMovie);
 
-        moviesTable.getItems().clear();
-        moviesTable.setItems(categoryModel.getAllMoviesFromCategoriesObservable(selectedCategory));
+        if (selectedMovie != null && selectedCategory != null){
+            ObservableList<Movie> moviesInCategory = categoryModel.getAllMoviesFromCategoriesObservable(selectedCategory);
+            for(Movie movie : moviesInCategory){
+                if (movie.getId() == selectedMovie.getId()){
+                    alReadyInCategory = true;
+                    break;
+                }
+            }
+
+            if (!alReadyInCategory){
+                categoryModel.addMovieToCategory(selectedCategory, selectedMovie);
+                moviesInCategory.add(selectedMovie);
+                moviesTable.getItems().clear();
+                moviesTable.setItems(moviesInCategory);
+            }else{
+                error(selectedMovie.getName() +" already exists in "+ selectedCategory.getName());
+            }
+        }else {
+            error("Select a movie and which category you wish to add it to");
+        }
+
     }
 
-
     public void handleRadioButton(ActionEvent actionEvent) {
-        if (radioButtonTitle.isSelected()){
+        if (radioButtonTitle.isSelected()) {
             filterType = "movieFilter";
-        } else if(radioButtonRating.isSelected()){
+        } else if (radioButtonRating.isSelected()) {
             filterType = "ratingFilter";
-        }else if (radioButtonCategory.isSelected()){
+        } else if (radioButtonCategory.isSelected()) {
             filterType = "categoryFilter";
         }
     }
 
     /**
      * handleButtonPlay Styrer Play/Pause stop knappen, som enten starter eller stopper en film i at blive afspillet.
+     *
      * @param actionEvent javaFX event klasse
      * @throws IOException
      * @throws SQLServerException
      */
-
     public void handleButtonPlay(ActionEvent actionEvent) throws IOException, SQLServerException {
+        ///TODO UPDATE LAST VIEW NÅR MOVIE AFSPILLES FRA CATEGORY, IKKE KUN FRA ALLMOVIES TABLE
         if (allMoviesTable.getSelectionModel().getSelectedItem() != null) {
             Movie movie = allMoviesTable.getSelectionModel().getSelectedItem();
             movieModel.updateMovieTime(movie);
@@ -291,27 +322,37 @@ public class MainViewController implements Initializable {
             Desktop.getDesktop().open(new File(moviePath + movie2.getFileLink()));
             moviesTable.refresh();
         } else
-            error("Please select a movie");
-
+            error("Please select a movie to play");
     }
 
     public void lookAtCategoryMovies() {
-        Category category = categoryListView.getSelectionModel().getSelectedItem();
-        ObservableList<Movie> observableList = categoryModel.getAllMoviesFromCategoriesObservable(category);
-        moviesTable.setItems(observableList);
+        if (categoryListView.getSelectionModel().getSelectedItem() != null){
+            Category category = categoryListView.getSelectionModel().getSelectedItem();
+            ObservableList<Movie> observableList = categoryModel.getAllMoviesFromCategoriesObservable(category);
+            moviesTable.setItems(observableList);
+        }
     }
 
     public void deleteMovieFromCategory(ActionEvent actionEvent) {
+
         Category selectedCategory = categoryListView.getSelectionModel().getSelectedItem();
         Movie selectedMovie = moviesTable.getSelectionModel().getSelectedItem();
-        categoryModel.deleteMovieFromCategory(selectedCategory, selectedMovie);
+        if (selectedCategory != null && selectedMovie != null){
+            categoryModel.deleteMovieFromCategory(selectedCategory, selectedMovie);
 
-        moviesTable.getItems().clear();
-        moviesTable.setItems(categoryModel.getAllMoviesFromCategoriesObservable(selectedCategory));
+            moviesTable.getItems().clear();
+            moviesTable.setItems(categoryModel.getAllMoviesFromCategoriesObservable(selectedCategory));
+        }
+        else{
+            error("Select a category and what movie you wish to delete");
+        }
     }
 
     public void handleMovieTableClicked(MouseEvent mouseEvent) {
         allMoviesTable.getSelectionModel().clearSelection();
+    }
+    public void allMoviesTableClicked(MouseEvent mouseEvent) {
+        moviesTable.getSelectionModel().clearSelection();
     }
 
     public void playMovieMediaView(ActionEvent actionEvent) throws IOException {
@@ -334,11 +375,8 @@ public class MainViewController implements Initializable {
         }
     }
 
-    public void allMoviesTableClicked(MouseEvent mouseEvent) {
-    }
-
     private void oldMoviesWarning() throws IOException {
-        if (!movieModel.checkForOldMovies().isEmpty()){
+        if (!movieModel.checkForOldMovies().isEmpty()) {
             FXMLLoader root = new FXMLLoader(getClass().getResource("/gui/view/OldMoviesView.fxml"));
             Scene mainWindowScene = new Scene(root.load());
